chore/#39: extract sidebar to its own component.
